+++ src/org/usfirst/frc/team246/robot/OverclockedLibraries/CANTalonPotentiometer.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team246.robot;
+package org.usfirst.frc.team246.robot.OverclockedLibraries;
 
 import edu.wpi.first.wpilibj.CANTalon;
