+++ src/org/usfirst/frc/team246/robot/OverclockedLibraries/VectorPIDOutput.java
@@ -1,7 +1,5 @@
 package org.usfirst.frc.team246.robot.OverclockedLibraries;
 
-import org.usfirst.frc.team246.robot.Vector2D;
-
 public interface VectorPIDOutput {
 	public void pidWrite(Vector2D vector);
 }
