+++ src/org/usfirst/frc/team246/robot/OverclockedLibraries/VectorPIDSource.java
@@ -1,7 +1,5 @@
 package org.usfirst.frc.team246.robot.OverclockedLibraries;
 
-import org.usfirst.frc.team246.robot.Vector2D;
-
 public interface VectorPIDSource {
 	public Vector2D pidGet();
 }
