+++ src/org/usfirst/frc/team246/robot/commands/AutoDrive.java
@@ -1,7 +1,7 @@
 package org.usfirst.frc.team246.robot.commands;
 
 import org.usfirst.frc.team246.robot.Robot;
-import org.usfirst.frc.team246.robot.Vector2D;
+import org.usfirst.frc.team246.robot.OverclockedLibraries.Vector2D;
 
 import edu.wpi.first.wpilibj.command.Command;
