+++ src/org/usfirst/frc/team246/robot/commands/DrivingCommand.java
@@ -7,7 +7,7 @@ package org.usfirst.frc.team246.robot.commands;
 import edu.wpi.first.wpilibj.command.Command;
 
 import org.usfirst.frc.team246.robot.Robot;
-import org.usfirst.frc.team246.robot.Vector2D;
+import org.usfirst.frc.team246.robot.OverclockedLibraries.Vector2D;
 
 /**
  * An abstract subclass of CommandBase used as a basis for all other commands
